pyproject.toml
[project]
name = "omni"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=1.1.0",
    "keyboard>=0.13.5",
    "markdown>=3.10.2",
    "openpyxl>=3.1.5",
    "pyaudio>=0.2.14",
    "pyaudiowpatch>=0.2.12.8",
    "pygments>=2.19.2",
    "pyinstaller>=6.19.0",
    "pypdf2>=3.0.1",
    "pyqt6>=6.10.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "speechrecognition>=3.14.6",
    "supabase>=2.28.0",
]
